!refactor(payment status): add cron jobs
In the first instace accessors had to manage
many queries to obtain if a subscription is valid
or not. Now I implemented a better way to improve that.
Cron Jobs. Every day the chron job checks if a PAID payment
is a expiration_date expirated lmao. In that way, if it's not
the status FIELD will be converted from PAGADO to VENCIDO.